packages/pillbug-websocket/pyproject.toml
[project]
name = "pillbug-websocket"
version = "0.1.0"
description = "Socket.IO websocket channel plugin for Pillbug"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillbug",
    "pydantic-settings>=2.11.0",
    "python-socketio>=5.11.0",
    "uvicorn>=0.35.0",
]

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[tool.uv.sources]
pillbug = { workspace = true }

[tool.uv.build-backend]
module-name = "pillbug_websocket"
module-root = "src"
